cast to float

+++ pennylane/labs/trotter_error/realspace/realspace_coefficients.py
@@ -612,6 +612,6 @@ def _numpy_to_dict(arr, threshold):
     for index in zip(*nz):
         if abs(arr[index]) > threshold:
             index = tuple(map(int, index))
-            d[index] = int(arr[index])
+            d[index] = float(arr[index])
 
     return d
